elimina viewset.py
